remove Shrike support
Shrike is no longer used. Let's remove it to tighten up what's needed in
reporting functionality.

Note this requires a database migration. It should be the antithesis of
f111620bb8 which was added in add_shrike_and_parent_id_columns.py, but
there's a bug in the downgrade logic in that revision - "parent_sid" not
"parent_id".